feat: added todo list UI
